ft_getchar.c, ft_bzero.c and ft_strchar.c added

+++ ft_memset.c
@@ -6,20 +6,20 @@
 /*   By: sminnaar <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/19 14:44:32 by sminnaar          #+#    #+#             */
-/*   Updated: 2019/05/23 09:50:32 by sminnaar         ###   ########.fr       */
+/*   Updated: 2019/05/27 11:17:17 by sminnaar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memset(void *b, int c, unsigned long len)
+void	*ft_memset(void *b, int c, size_t len)
 {
 	char *a;
 
 	a = (char *)b;
 	while (len > 0)
 	{
-		a[len - 1] = (unsigned char )c;
+		a[len - 1] = (unsigned char)c;
 		len--;
 	}
 	return (b);
