+++ ft_strcmp.c
@@ -6,7 +6,7 @@
 /*   By: sminnaar <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/19 15:02:13 by sminnaar          #+#    #+#             */
-/*   Updated: 2019/05/27 10:54:08 by sminnaar         ###   ########.fr       */
+/*   Updated: 2019/05/27 11:18:08 by sminnaar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strcmp(const char *s1, const char *s2)
 {
-	unsigned int i;
+	size_t i;
 
 	i = 0;
 	if (!s1 || !s2)
