+++ ft_strnequ.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int	ft_strnequ(char const *s1, char const *s2, unsigned long n)
+int	ft_strnequ(char const *s1, char const *s2, size_t n)
 {
 	if (n == 0)
 		return (1);
